# ssh: optimize detection