p10k's ssh detection is pretty thorough, but also pretty heavy, especially when the primary use case is local. on my m1 macbook, `(( $+commands[who] )) && return` costs 7 milliseconds.

this PR changes `$+commands[who]` to `command -v who`, which runs almost instantaneously in comparison.

the next biggest cost is in the regex match at the end of the function. it feels like ssh detection should be an optional thing. the best way i've found to bypass it is to set `P9K_SSH` and faking out `_P9K_TTY`, and forcing detection on `SSH_CLIENT` or `SSH_CONNECTION` alone, which i'm fine with.