@Syphdias 2024-10-16
How did you benchmark that the new comparison is faster? And how big is `commands` for you?