@Syphdias 2024-10-16
I had a hard time believing it and needed to give this a try myself. Note I am no Linux, but this shouldn't be that different on an M1
```zsh
#!/usr/bin/env zsh
COUNT=${COUNT:-10000}
function old() {
    (( $+commands[who] )) || return
}
function new() {
    [[ -z "$(command -v who)" ]] && return
}

t0=$(date +%s)

for i in {0..$COUNT}; do
    old
done

t1=$(date +%s)

for i in {0..$COUNT}; do
    new
done

t2=$(date +%s)

for i in {0..$COUNT}; do
    (( $+commands[who] ))
done

t3=$(date +%s)

for i in {0..$COUNT}; do
    [[ -z "$(command -v who)" ]]
done

t4=$(date +%s)

{
    echo "command total average"
    echo "+commandsF: $((t1-t0))s $(bc -l <<< "scale=5; 1000*($t1-$t0)/$COUNT")ms"
    echo "command-vF: $((t2-t1))s $(bc -l <<< "scale=5; 1000*($t2-$t1)/$COUNT")ms"
    echo "+commands: $((t3-t2))s $(bc -l <<< "scale=5; 1000*($t3-$t2)/$COUNT")ms"
    echo "command-v: $((t4-t3))s $(bc -l <<< "scale=5; 1000*($t4-$t3)/$COUNT")ms"
} |column -t
```
```
❯ COUNT=100000 /tmp/commands
command      total  average
+commandsF:  1s     .01000ms
command-vF:  72s    .72000ms
+commands:   0s     0ms
command-v:   49s    .49000ms
```

Note that this measuring is very inaccurate since I call `date` to measure points in time. I also tried them inside and outside functions. In this case any inaccuracy does not matter. ` (( $+commands[who] ))` is leagues faster than ` [[ -z "$(command -v who)" ]]`…