@z0rc 2024-10-16
@Syphdias instead of `date` you can use `$EPOCHREALTIME` from `zsh/datetime` module https://zsh.sourceforge.io/Doc/Release/Zsh-Modules.html#index-EPOCHREALTIMEю