@Syphdias 2024-10-17
> My favorite recipe for quick-n-dirty benchmarks:

I used that first with `{}` instead of `()` but wanted to make sure I tested in fair conditions in case I missed something. I thought `$()` would always fork but I read somewhere that is not the case for simple commands, but I did not trust that. Maybe this optimisation was only for scripts or something.

Any how, as expected, searching through a variable is faster than calling a program/function…