@visigoth 2024-10-29
> How did you benchmark that the new comparison is faster? And how big is `commands` for you?

i used hyperfine to run the command inside a terminal. i don't have the command handy in my terminal history at the moment, though. for reference `echo $#commands` prints 4009.